gradle 9.7.0, 5-os CI matrix, detekt, zero compiler warnings
- gradle wrapper 9.6.1 -> 9.7.0
- ci.yml: matrix linux/macos amd64+arm64 + windows-amd64, Liberica JDK 25,
  bash shells for portable steps
- detekt 1.23.8 wired into check with parity-friendly config; JVM-25
  workarounds (jvmTarget=21 pin + java.version spoof during analysis)
- warnings to zero: JvmOverloads removed (no default args), deprecated
  ObjectNode.put -> replace, dead null-condition dropped
  (SurgicalPatchTest), redundant toInt/toLong conversions removed
  (AssetBundleFile, AssetsFile, TypeTreeType)

+++ assetapi/src/main/kotlin/com/github/jpabscale/asset4j/bundle/AssetBundleFile.kt
@@ -475,7 +475,7 @@ class AssetBundleFile {
         if (newHeader.version >= 7)
             writer.align16()
 
-        val headerSize = (writer.position - startPos).toInt()
+        val headerSize = writer.position - startPos
 
         var totalCompressedSize = 0L
         val newBlocks = mutableListOf<AssetBundleBlockInfo>()
@@ -541,7 +541,7 @@ class AssetBundleFile {
         val bundleInfoBytes = infoWriter.toByteArray()
         val bundleInfoBytesCom = if (compType == AssetBundleCompressionType.LZ4Fast) UnityCompression.compressLz4Fast(bundleInfoBytes) else UnityCompression.compressLz4(bundleInfoBytes)
 
-        val totalFileSize = (headerSize + bundleInfoBytesCom.size + totalCompressedSize).toLong()
+        val totalFileSize = headerSize + bundleInfoBytesCom.size + totalCompressedSize
         newFsHeader.totalFileSize = totalFileSize
         newFsHeader.decompressedSize = bundleInfoBytes.size.toLong()
         newFsHeader.compressedSize = bundleInfoBytesCom.size.toLong()
